+++ hive/lib/src/backend/storage_backend_vm.dart
@@ -8,9 +8,9 @@ import 'package:hive/src/binary/binary_writer_impl.dart';
 import 'package:hive/src/binary/frame.dart';
 import 'package:hive/src/box/box_base.dart';
 import 'package:hive/src/box/box_options.dart';
-import 'package:hive/src/box/cached_box.dart';
+import 'package:hive/src/box/cached_box_impl.dart';
 import 'package:hive/src/box/keystore.dart';
-import 'package:hive/src/box/lazy_box.dart';
+import 'package:hive/src/box/lazy_box_impl.dart';
 import 'package:hive/src/crypto_helper.dart';
 import 'package:hive/src/hive_impl.dart';
 import 'package:hive/src/io/buffered_file_reader.dart';
@@ -19,7 +19,8 @@ import 'package:hive/src/io/synced_file.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 
-Future<Box> openBox(HiveImpl hive, String name, BoxOptions options) async {
+Future<Box> openBox(
+    HiveImpl hive, String name, bool lazy, BoxOptions options) async {
   var file = await findHiveFileAndCleanUp(name, hive.path);
 
   CryptoHelper crypto;
@@ -31,10 +32,10 @@ Future<Box> openBox(HiveImpl hive, String name, BoxOptions options) async {
 
   var backend = StorageBackendVm(syncedFile, crypto);
   BoxBase box;
-  if (options.lazy) {
-    box = LazyBox(hive, name, options, backend);
+  if (lazy) {
+    box = LazyBoxImpl(hive, name, options, backend);
   } else {
-    box = CachedBox(hive, name, options, backend);
+    box = CachedBoxImpl(hive, name, options, backend);
   }
   backend._registry = box;
 
@@ -94,7 +95,7 @@ class StorageBackendVm extends StorageBackend {
 
   @override
   Future<int> initialize(
-      Map<String, BoxEntry> entries, bool lazy, bool crashRecovery) async {
+      Map<dynamic, BoxEntry> entries, bool lazy, bool crashRecovery) async {
     var frames = <Frame>[];
     int recoveryOffset;
     if (!lazy) {
@@ -135,20 +136,20 @@ class StorageBackendVm extends StorageBackend {
   }
 
   @override
-  Future<dynamic> readValue(String key, int offset, int length) async {
+  Future<dynamic> readValue(dynamic key, int offset, int length) async {
     var bytes = await _file.readAt(offset, length);
     var frame = Frame.fromBytes(bytes, _registry, _crypto);
     return frame.value;
   }
 
   @override
-  Future<Map<String, dynamic>> readAll() async {
+  Future<Map<dynamic, dynamic>> readAll() async {
     var frames = <Frame>[];
     await _file.writeLock.synchronized(() {
       return _helper.readFramesFromFile(path, frames, _registry, _crypto);
     });
 
-    var map = <String, dynamic>{};
+    var map = <dynamic, dynamic>{};
     for (var frame in frames) {
       map[frame.key] = frame.value;
     }
@@ -185,11 +186,11 @@ class StorageBackendVm extends StorageBackend {
   }
 
   @override
-  Future<Map<String, BoxEntry>> compact(Map<String, BoxEntry> entries) async {
+  Future<Map<dynamic, BoxEntry>> compact(Map<dynamic, BoxEntry> entries) async {
     var compactPath = '${p.withoutExtension(path)}.hivec';
     var compactFile = await File(compactPath).open(mode: FileMode.write);
 
-    var newEntries = HashMap<String, BoxEntry>();
+    var newEntries = HashMap<dynamic, BoxEntry>();
 
     await _file.readLock.synchronized(() {
       return _file.writeLock.synchronized(() async {