+++ hive/lib/src/box/box_options.dart
@@ -2,13 +2,11 @@ import 'package:hive/hive.dart';
 
 class BoxOptions {
   final List<int> encryptionKey;
-  final bool lazy;
   final CompactionStrategy compactionStrategy;
   final bool crashRecovery;
 
   BoxOptions({
     this.encryptionKey,
-    this.lazy,
     this.compactionStrategy,
     this.crashRecovery,
   });