+++ hive/CHANGELOG.md
@@ -1,11 +1,15 @@
 ## 0.4.0
-- Added []= operator to boxes
+- Added []= operator to lazy boxes
 - Added `BigInt` support
 - Added `compactionStrategy` parameter
 - Added automatic crash recovery
+- Added `add()` and `addAll()` for auto increment keys
+- Support for int (32 bit unsigned) keys
 - Cached boxes now notify their listeners immediately about changes
 - Bugfixes
 - More tests
+- **Breaking:** Binary format changed
+- **Breaking:** Minor API changes
 
 ## 0.3.0+1
 - Bugfix: `Hive['yourBox']` didn't work with uppercase box names